requirements.txt
Django
django-ninja
django-ninja-jwt[crypto]
psycopg[binary]
python-decouple
pydantic
python-dotenv
celery
redis
django-celery-results
pytest
pytest-django
python-docx
pyjwt
openai
dj_database_url
requests
rav
psycopg2-binary
rest_framework_simplejwt
python-zarinpal
django-zarinpal
